Move `get_nested_result` into new public `helpers` module.
